Move counter -> App for svelte projects
