+++ frameworks/svelte/incrementing-render-effect/src/App.svelte
@@ -1,7 +1,21 @@
 <script lang="ts">
-  import Counter from './lib/Counter.svelte'
+  import { helpers } from 'common';
+
+  const test = helpers.incrementingRenderEffect();
+  let output: number = $state(0);
+  let advancer: (() => void) | undefined = $state();
+
+  function run() {
+    let value = output;
+    advancer?.();
+    return value;
+  }
+
+  test.doit({
+    get: () => output,
+    set: (value: number) => output = value,
+    setupAdvancer: (fn: () => void) => { advancer = fn; },
+  });
 </script>
 
-<main>
-    <Counter />
-</main>
+<output>{run()}</output>
