--- frameworks/svelte/one-item-many-updates/src/lib/Counter.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-  import { helpers } from 'common';
-
-  let test = helpers.oneItem10kUpdates();
-  let count: number = $state(test.getData())
-
-  $effect(() => {
-    test.doit((i) => count = i);
-  });
-
-  
-</script>
-
-<output>{test.formatItem(count)}</output>
