--- frameworks/svelte/ten-k-items-one-time/src/lib/Counter.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-  import { helpers } from 'common';
-
-  let test = helpers.tenKitems1UpdateEach();
-  let items = $state(test.getData());
-
-  $effect(() => {
-    test.doit((i) => {
-      items[i] = i;
-    });
-  });
-</script>
-
-{#each items as item}{test.formatItem(item)}{/each}
